+++ src/main/java/subway/line/controller/LineController.java
@@ -5,10 +5,10 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import subway.line.domain.Line;
-import subway.line.domain.LineRequest;
-import subway.line.domain.LineResponse;
+import subway.line.dto.LineRequest;
+import subway.line.dto.LineResponse;
 import subway.line.service.LineService;
-import subway.section.domain.SectionRequest;
+import subway.section.dto.SectionRequest;
 import subway.section.domain.Sections;
 import subway.section.service.SectionService;
 import subway.station.domain.Stations;
