+++ src/main/java/subway/section/service/SectionService.java
@@ -3,9 +3,9 @@ package subway.section.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import subway.exception.EntityNotFoundException;
-import subway.line.domain.LineRequest;
+import subway.line.dto.LineRequest;
 import subway.section.dao.SectionDao;
-import subway.section.domain.SectionRequest;
+import subway.section.dto.SectionRequest;
 import subway.section.domain.Sections;
 import subway.station.dao.StationDao;
 import subway.station.domain.Stations;
