+++ src/test/java/subway/station/StationAcceptanceTest.java
@@ -8,8 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import subway.AcceptanceTest;
-import subway.station.domain.StationRequest;
-import subway.station.domain.StationResponse;
+import subway.station.dto.StationRequest;
+import subway.station.dto.StationResponse;
 
 import java.util.Arrays;
 import java.util.List;
